+++ react/Summary.tsx
@@ -1,6 +1,7 @@
 import React, { FC } from 'react'
 import { ExtensionPoint } from 'vtex.render-runtime'
-import { useOrderForm } from 'vtex.order-manager/OrderForm'
+import { OrderForm as OrderFormComponent } from 'vtex.order-manager'
+import { OrderForm } from 'vtex.checkout-graphql'
 import { useCssHandles, CssHandlesTypes } from 'vtex.css-handles'
 
 const CSS_HANDLES = ['minicartSummary'] as const
@@ -10,11 +11,19 @@ interface Props {
 }
 
 const Summary: FC<Props> = ({ classes }) => {
+  const { useOrderForm } = OrderFormComponent
+
   const {
-    orderForm: { totalizers, value, paymentData },
-  } = useOrderForm()
+    orderForm: { totalizers, value, paymentData, items = [] },
+  }: { orderForm: OrderForm } = useOrderForm()
 
   const { handles } = useCssHandles(CSS_HANDLES, { classes })
+  const originalTotal = items.reduce(
+    (total, item) =>
+      (total as number) +
+      ((item?.listPrice as number) ?? 0) * (item?.quantity ?? 1),
+    0
+  )
 
   return (
     <div className={`${handles.minicartSummary} ph4 ph6-l pt5`}>
@@ -23,6 +32,7 @@ const Summary: FC<Props> = ({ classes }) => {
         totalizers={totalizers}
         paymentData={paymentData}
         total={value}
+        originalTotal={originalTotal}
       />
     </div>
   )